refactor with zero functional or logic changes
