+++ .gitignore
@@ -5,3 +5,4 @@ emsys
 unicodetest
 /build
 /config.h
+/tests/test_example
