+++ fileio.c
@@ -1,0 +1,136 @@
+#define _DEFAULT_SOURCE
+#define _BSD_SOURCE
+#define _GNU_SOURCE
+
+#include "emsys.h"
+#include "fileio.h"
+#include "buffer.h"
+#include "tab.h"
+#include <errno.h>
+#include <fcntl.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include "display.h"
+
+/* Access global editor state */
+extern struct editorConfig E;
+
+/* External functions we need */
+extern void die(const char *s);
+extern char *stringdup(const char *s);
+
+/*** file i/o ***/
+
+char *editorRowsToString(struct editorBuffer *bufr, int *buflen) {
+	int totlen = 0;
+	int j;
+	for (j = 0; j < bufr->numrows; j++) {
+		totlen += bufr->row[j].size + 1;
+	}
+	*buflen = totlen;
+
+	char *buf = malloc(totlen);
+	char *p = buf;
+	for (j = 0; j < bufr->numrows; j++) {
+		memcpy(p, bufr->row[j].chars, bufr->row[j].size);
+		p += bufr->row[j].size;
+		*p = '\n';
+		p++;
+	}
+
+	return buf;
+}
+
+void editorOpen(struct editorBuffer *bufr, char *filename) {
+	free(bufr->filename);
+	bufr->filename = stringdup(filename);
+	FILE *fp = fopen(filename, "r");
+	if (!fp) {
+		if (errno == ENOENT) {
+			editorSetStatusMessage("(New file)", bufr->filename);
+			return;
+		}
+		die("fopen");
+	}
+
+	char *line = NULL;
+	size_t linecap = 0;
+	ssize_t linelen;
+	/* Doesn't handle null bytes */
+	while ((linelen = getline(&line, &linecap, fp)) != -1) {
+		while (linelen > 0 &&
+		       (line[linelen - 1] == '\n' || line[linelen - 1] == '\r'))
+			linelen--;
+		editorInsertRow(bufr, bufr->numrows, line, linelen);
+	}
+	free(line);
+	fclose(fp);
+	bufr->dirty = 0;
+}
+
+void editorRevert(struct editorConfig *ed, struct editorBuffer *buf) {
+	struct editorBuffer *new = newBuffer();
+	editorOpen(new, buf->filename);
+	new->next = buf->next;
+	ed->focusBuf = new;
+	if (ed->firstBuf == buf) {
+		ed->firstBuf = new;
+	}
+	struct editorBuffer *cur = ed->firstBuf;
+	while (cur != NULL) {
+		if (cur->next == buf) {
+			cur->next = new;
+			break;
+		}
+		cur = cur->next;
+	}
+	for (int i = 0; i < ed->nwindows; i++) {
+		if (ed->windows[i]->buf == buf) {
+			ed->windows[i]->buf = new;
+		}
+	}
+	new->indent = buf->indent;
+	new->cx = buf->cx;
+	new->cy = buf->cy;
+	if (new->cy > new->numrows) {
+		new->cy = new->numrows;
+		new->cx = 0;
+	} else if (new->cx > new->row[new->cy].size) {
+		new->cx = new->row[new->cy].size;
+	}
+	destroyBuffer(buf);
+}
+
+void editorSave(struct editorBuffer *bufr) {
+	if (bufr->filename == NULL) {
+		bufr->filename = (char *)
+			editorPrompt(bufr, (uint8_t *)"Save as: %s", PROMPT_FILES, NULL);
+		if (bufr->filename == NULL) {
+			editorSetStatusMessage("Save aborted.");
+			return;
+		}
+	}
+
+	int len;
+	char *buf = editorRowsToString(bufr, &len);
+
+	int fd = open(bufr->filename, O_RDWR | O_CREAT, 0644);
+	if (fd != -1) {
+		if (ftruncate(fd, len) != -1) {
+			if (write(fd, buf, len)) {
+				close(fd);
+				free(buf);
+				bufr->dirty = 0;
+				editorSetStatusMessage("Wrote %d bytes to %s",
+						       len, bufr->filename);
+				return;
+			}
+		}
+		close(fd);
+	}
+
+	free(buf);
+	editorSetStatusMessage("Save failed: %s", strerror(errno));
+}
