+++ terminal.c
@@ -1,0 +1,91 @@
+#include "terminal.h"
+#include "emsys.h"
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <sys/ioctl.h>
+#include <termios.h>
+#include <unistd.h>
+#include <string.h>
+#include "unicode.h"
+
+extern struct editorConfig E;
+
+void die(const char *s) {
+	write(STDOUT_FILENO, CSI "2J", 4);
+	write(STDOUT_FILENO, CSI "H", 3);
+	perror(s);
+	write(STDOUT_FILENO, CRLF, 2);
+	write(STDOUT_FILENO, "sleeping 5s", 11);
+	sleep(5);
+	exit(1);
+}
+
+void disableRawMode() {
+	if (tcsetattr(STDIN_FILENO, TCSAFLUSH, &E.orig_termios) == -1)
+		die("disableRawMode tcsetattr");
+	if (write(STDOUT_FILENO, CSI "?1049l", 8) == -1)
+		die("disableRawMode write");
+}
+
+void enableRawMode() {
+	/* Saves the screen and switches to an alt screen */
+	if (write(STDOUT_FILENO, CSI "?1049h", 8) == -1)
+		die("enableRawMode write");
+
+	/*
+	 * I looked into it. It's possible, but not easy, to do it
+	 * without termios. Basically you'd have to hand-hack and send
+	 * off your own bits. Check out busybox vi and that rabbithole
+	 * for an implementation.
+	 */
+	if (tcgetattr(STDIN_FILENO, &E.orig_termios) == -1)
+		die("tcgetattr");
+	atexit(disableRawMode);
+
+	struct termios raw = E.orig_termios;
+	raw.c_iflag &= ~(BRKINT | ICRNL | INPCK | ISTRIP | IXON);
+	raw.c_oflag &= ~(OPOST);
+	raw.c_cflag |= (CS8);
+	raw.c_lflag &= ~(ECHO | ICANON | IEXTEN | ISIG);
+	if (tcsetattr(STDIN_FILENO, TCSAFLUSH, &raw) == -1)
+		die("enableRawMode tcsetattr");
+}
+
+int getCursorPosition(int *rows, int *cols) {
+	char buf[32];
+	int i = 0;
+
+	if (write(STDOUT_FILENO, "\x1b[6n", 4) != 4)
+		return -1;
+
+	while (i < (unsigned int)(sizeof(buf) - 1)) {
+		if (read(STDIN_FILENO, &buf[i], 1) != 1)
+			break;
+		if (buf[i] == 'R')
+			break;
+		i++;
+	}
+	buf[i] = '\0';
+
+	if (buf[0] != '\x1b' || buf[1] != '[')
+		return -1;
+	if (sscanf(&buf[2], "%d;%d", rows, cols) != 2)
+		return -1;
+
+	return 0;
+}
+
+int getWindowSize(int *rows, int *cols) {
+	struct winsize ws;
+
+	if (ioctl(STDOUT_FILENO, TIOCGWINSZ, &ws) == -1 || ws.ws_col == 0) {
+		if (write(STDOUT_FILENO, "\x1b[999C\x1b[999B", 12) != 12)
+			return -1;
+		return getCursorPosition(rows, cols);
+	} else {
+		*cols = ws.ws_col;
+		*rows = ws.ws_row;
+		return 0;
+	}
+}
