+++ transform.c
@@ -1,9 +1,15 @@
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
-#include "bound.h"
 #include "transform.h"
+#include "edit.h"
 #include "unicode.h"
+#include "emsys.h"
+#include "buffer.h"
+#include "undo.h"
+#include "display.h"
+#include "unused.h"
+#include "region.h"
 
 #define MKOUTPUT(in, l, o)  \
 	int l = strlen(in); \
@@ -100,4 +106,37 @@ uint8_t *transformerTransposeWords(uint8_t *input) {
 	output[len] = 0;
 
 	return output;
+}
+
+void editorCapitalizeRegion(struct editorConfig *ed, struct editorBuffer *buf) {
+	editorTransformRegion(ed, buf, transformerCapitalCase);
+}
+
+void editorWhitespaceCleanup(struct editorConfig *UNUSED(ed),
+			     struct editorBuffer *buf) {
+	unsigned int trailing = 0;
+	for (int i = 0; i < buf->numrows; i++) {
+		erow *row = &buf->row[i];
+		for (int j = row->size - 1; j >= 0; j--) {
+			if (row->chars[j] == ' ' || row->chars[j] == '\t') {
+				row->size--;
+				trailing++;
+			} else {
+				break;
+			}
+		}
+		editorUpdateRow(row);
+	}
+
+	if (buf->cx > buf->row[buf->cy].size) {
+		buf->cx = buf->row[buf->cy].size;
+	}
+
+	if (trailing > 0) {
+		clearUndosAndRedos(buf);
+		editorSetStatusMessage("%d trailing characters removed",
+				       trailing);
+	} else {
+		editorSetStatusMessage("No change.");
+	}
 }
